- removed .idea folder as that shouldn't have been uploaded in the first place ~ changed g_Time to float instead of double

+++ src/WallpaperEngine/Render/Objects/Effects/CPass.cpp
@@ -20,7 +20,7 @@ using namespace WallpaperEngine::Render::Shaders::Variables;
 
 using namespace WallpaperEngine::Render::Objects::Effects;
 
-extern double g_Time;
+extern float g_Time;
 
 CPass::CPass (CMaterial* material, Core::Objects::Images::Materials::CPass* pass) :
     m_material (material),
